Corrected compilation structure for rnprotocol dir
Removed the rnprotocol Makefile and adjusted compilation structure.

Originally both rnprotocol, servercomm and clientcomm directories
were pushing files into libservercomm.a and libclientcomm.a.
Because the libraries stay in the directories where they were created
the libs were never merged.
While this was okay for the server it is impossible to deploy both
libraries that contain all the required object files.

Now the servercomm and cliencomm are responsible for compiling the
files in rnprotocol and joining them into the respective libraries.