Remove EARLY_TEMPLATE from rasql_signal
The signal handlers don't use any templated classes so there is no need to
pull in this header. Also, using this header can lead to templated classes
being defined in both `rasql.o` and `rasql_signal.o` after compilation which
causes the linker to abort with a duplicate symbol error.